requirements.txt
python-telegram-bot[webhooks]>=21.0,<22.0
httpx>=0.26.0
python-dotenv>=1.0.1
